refactor(tests): C8-02/03/05/08 relocate misplaced test coverage
Move tests to their canonical homes so each test file matches the module
it exercises:

- C8-02: test_path_mtime_ns -> tests/recipe/test__api.py (from test_io.py)
- C8-03: TestReadDb + routing test -> tests/server/test_tools_status.py
  (from test_tools_workspace.py; tools_status owns read_db)
- C8-05: Add tests/docs/ entry to tests/CLAUDE.md directory tree
- C8-08 (partial): Move report_bug, fetch_github_issue, get_issue_title
  self-contained unit tests from test_tools_integrations.py to
  test_tools_github.py; drop duplicates already covered in target.

Also restore missing `import pytest` in tests/server/test_factory.py
(collection error from C2-15 parametrize refactor), and pick up ruff
format/unused-import cleanups auto-applied by pre-commit.
