+++ tests/recipe/test__api.py
@@ -29,3 +29,18 @@ def test_drop_sub_recipe_step_preserves_future_fields() -> None:
     assert result.kitchen_rules == ["rule1"]
     assert "placeholder" not in result.steps
     assert "other" in result.steps
+
+
+# ---------------------------------------------------------------------------
+# T2 — _path_mtime_ns replaces the two old helpers
+# (migrated from recipe/test_io.py — this test introspects recipe._api)
+# ---------------------------------------------------------------------------
+
+
+def test_path_mtime_ns_exists_and_old_helpers_removed() -> None:
+    """recipe/_api.py must expose _path_mtime_ns; _file_mtime_ns/_dir_mtime_ns removed."""
+    import autoskillit.recipe._api as api
+
+    assert hasattr(api, "_path_mtime_ns"), "_path_mtime_ns must exist"
+    assert not hasattr(api, "_file_mtime_ns"), "_file_mtime_ns must be removed"
+    assert not hasattr(api, "_dir_mtime_ns"), "_dir_mtime_ns must be removed"
