+++ tests/server/test_tools_workspace.py
@@ -4,18 +4,15 @@ from __future__ import annotations
 
 import json
 import os
-from unittest.mock import MagicMock
 
 import pytest
 
 from autoskillit.config import (
     AutomationConfig,
-    ReadDbConfig,
     ResetWorkspaceConfig,
     SafetyConfig,
 )
 from autoskillit.core.types import AUTOSKILLIT_PRIVATE_ENV_VARS
-from autoskillit.server.tools_status import read_db
 from autoskillit.server.tools_workspace import reset_test_dir, reset_workspace, test_check
 from autoskillit.workspace import CleanupResult
 from tests.conftest import _make_result
@@ -386,224 +383,6 @@ class TestConfigDefaults:
     def test_default_classify_fix_empty_prefixes(self):
         cfg = AutomationConfig()
         assert cfg.classify_fix.path_prefixes == []
-
-
-@pytest.mark.usefixtures("tool_ctx")
-class TestReadDb:
-    """Integration tests for read_db tool with real SQLite databases."""
-
-    @pytest.fixture
-    def sample_db(self, tmp_path):
-        """Create a sample SQLite database for testing."""
-        import sqlite3
-
-        db = tmp_path / "test.db"
-        conn = sqlite3.connect(str(db))
-        conn.execute("CREATE TABLE users (id INTEGER, name TEXT, age INTEGER)")
-        conn.execute("INSERT INTO users VALUES (1, 'Alice', 30)")
-        conn.execute("INSERT INTO users VALUES (2, 'Bob', 25)")
-        conn.execute("INSERT INTO users VALUES (3, 'Charlie', 35)")
-        conn.commit()
-        conn.close()
-        return db
-
-    @pytest.mark.anyio
-    async def test_simple_select(self, sample_db):
-        result = json.loads(await read_db(db_path=str(sample_db), query="SELECT * FROM users"))
-        assert result["row_count"] == 3
-        assert result["column_names"] == ["id", "name", "age"]
-        assert len(result["rows"]) == 3
-        assert result["truncated"] is False
-
-    @pytest.mark.anyio
-    async def test_parameterized_query(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT name FROM users WHERE age > ?",
-                params="[28]",
-            )
-        )
-        assert result["row_count"] == 2
-        names = [r["name"] for r in result["rows"]]
-        assert "Alice" in names
-        assert "Charlie" in names
-
-    @pytest.mark.anyio
-    async def test_named_params(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT name FROM users WHERE age = :age",
-                params='{"age": 25}',
-            )
-        )
-        assert result["row_count"] == 1
-        assert result["rows"][0]["name"] == "Bob"
-
-    @pytest.mark.anyio
-    async def test_empty_result(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT * FROM users WHERE age > 100",
-            )
-        )
-        assert result["row_count"] == 0
-        assert result["rows"] == []
-        assert result["column_names"] == ["id", "name", "age"]
-
-    @pytest.mark.anyio
-    async def test_rejects_insert(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="INSERT INTO users VALUES (4, 'Dave', 40)",
-            )
-        )
-        assert "error" in result
-        err_lower = result["error"].lower()
-        assert "forbidden" in err_lower or "select" in err_lower or "not authorized" in err_lower
-
-    @pytest.mark.anyio
-    async def test_rejects_drop(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="DROP TABLE users",
-            )
-        )
-        assert "error" in result
-
-    @pytest.mark.anyio
-    async def test_rejects_attach(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="ATTACH DATABASE ':memory:' AS other",
-            )
-        )
-        assert "error" in result
-
-    @pytest.mark.anyio
-    async def test_nonexistent_db(self, tmp_path):
-        result = json.loads(
-            await read_db(
-                db_path=str(tmp_path / "nonexistent.db"),
-                query="SELECT 1",
-            )
-        )
-        assert "error" in result
-        assert "does not exist" in result["error"] or "not found" in result["error"].lower()
-
-    @pytest.mark.anyio
-    async def test_not_a_file(self, tmp_path):
-        result = json.loads(
-            await read_db(
-                db_path=str(tmp_path),
-                query="SELECT 1",
-            )
-        )
-        assert "error" in result
-
-    @pytest.mark.anyio
-    async def test_invalid_params_json(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT * FROM users",
-                params="not json",
-            )
-        )
-        assert "error" in result
-        assert "params" in result["error"].lower()
-
-    @pytest.mark.anyio
-    async def test_gated_when_disabled(self, sample_db, tool_ctx):
-        from autoskillit.pipeline.gate import DefaultGateState
-
-        tool_ctx.gate = DefaultGateState(enabled=False)
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT 1",
-            )
-        )
-        assert result["success"] is False
-        assert result["is_error"] is True
-        assert "not enabled" in result["result"].lower()
-
-    @pytest.mark.anyio
-    async def test_max_rows_truncation(self, sample_db, tool_ctx):
-        tool_ctx.config = AutomationConfig(read_db=ReadDbConfig(max_rows=2))
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT * FROM users",
-            )
-        )
-        assert result["row_count"] == 2
-        assert result["truncated"] is True
-
-    @pytest.mark.anyio
-    async def test_blob_base64_encoding(self, tmp_path):
-        import base64
-        import sqlite3
-
-        db = tmp_path / "blob.db"
-        conn = sqlite3.connect(str(db))
-        conn.execute("CREATE TABLE data (id INTEGER, content BLOB)")
-        conn.execute("INSERT INTO data VALUES (1, ?)", (b"\x00\x01\x02\xff",))
-        conn.commit()
-        conn.close()
-        result = json.loads(
-            await read_db(
-                db_path=str(db),
-                query="SELECT * FROM data",
-            )
-        )
-        assert base64.b64decode(result["rows"][0]["content"]) == b"\x00\x01\x02\xff"
-
-    @pytest.mark.anyio
-    async def test_query_timeout(self, sample_db, tool_ctx):
-        tool_ctx.config = AutomationConfig(read_db=ReadDbConfig(timeout=1))
-        # Cross join 3 rows^18 = ~387 million rows — guaranteed to exceed 1s timeout
-        slow_query = "SELECT count(*) FROM " + ", ".join(f"users t{i}" for i in range(18))
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query=slow_query,
-            )
-        )
-        assert "error" in result
-        assert "timeout" in result["error"].lower()
-
-    @pytest.mark.anyio
-    async def test_sql_error_returns_error(self, sample_db):
-        result = json.loads(
-            await read_db(
-                db_path=str(sample_db),
-                query="SELECT nonexistent_column FROM users",
-            )
-        )
-        assert "error" in result
-
-
-@pytest.mark.anyio
-async def test_tools_status_routes_through_db_reader(tool_ctx, tmp_path) -> None:
-    """read_db routes through ctx.db_reader.query()."""
-    import sqlite3 as _sqlite3
-
-    tool_ctx.db_reader = MagicMock()
-    tool_ctx.db_reader.query.return_value = {"rows": [], "count": 0}
-
-    db_path = str(tmp_path / "test.db")
-    # Create an empty sqlite db so path-exists check passes
-    _sqlite3.connect(db_path).close()
-    await read_db(db_path, "SELECT 1")
-    tool_ctx.db_reader.query.assert_called_once()
-    call_kwargs = tool_ctx.db_reader.query.call_args
-    assert "SELECT 1" in str(call_kwargs)
 
 
 @pytest.mark.anyio
